menuUtils: Move the menu helpers into their own module
Moves moveItemInMenu out of dbusMenu.js into a module of its own,
unchanged, and imports it back under the name it already had.
dbusMenu.js pulls in popupMenu.js from the gnome-shell resources, which
a plain gjs process cannot resolve, so the helper could not be reached
from a test without a running shell.
